refactor: Navigation bar splited in two components
